+++ src/Engines/QueryBuilderEngine.php
@@ -126,19 +126,51 @@ class QueryBuilderEngine extends BaseEngine
      */
     public function filtering()
     {
+        $keyword = $this->request->keyword();
+
+        if ($this->isSmartSearch()) {
+            $this->smartGlobalSearch($keyword);
+
+            return;
+        }
+
+        $this->globalSearch($keyword);
+    }
+
+    /**
+     * Perform multi-term search by splitting keyword into
+     * individual words and searches for each of them.
+     *
+     * @param string $keyword
+     */
+    private function smartGlobalSearch($keyword)
+    {
+        $keywords = array_filter(explode(' ', $keyword));
+
+        foreach ($keywords as $keyword) {
+            $this->globalSearch($keyword);
+        }
+    }
+
+    /**
+     * Perform global search for the given keyword.
+     *
+     * @param string $keyword
+     */
+    private function globalSearch($keyword)
+    {
         $this->query->where(
-            function ($query) {
-                $globalKeyword = $this->request->keyword();
-                $queryBuilder  = $this->getQueryBuilder($query);
+            function ($query) use ($keyword) {
+                $queryBuilder = $this->getQueryBuilder($query);
 
                 foreach ($this->request->searchableColumnIndex() as $index) {
                     $columnName = $this->getColumnName($index);
-                    if ($this->isBlacklisted($columnName)) {
+                    if ($this->isBlacklisted($columnName) && ! $this->hasCustomFilter($columnName)) {
                         continue;
                     }
 
                     // check if custom column filtering is applied
-                    if (isset($this->columnDef['filter'][$columnName])) {
+                    if ($this->hasCustomFilter($columnName)) {
                         $columnDef = $this->columnDef['filter'][$columnName];
                         // check if global search should be applied for the specific column
                         $applyGlobalSearch = count($columnDef['parameters']) == 0 || end($columnDef['parameters']) !== false;
@@ -148,7 +180,7 @@ class QueryBuilderEngine extends BaseEngine
 
                         if ($columnDef['method'] instanceof Closure) {
                             $whereQuery = $queryBuilder->newQuery();
-                            call_user_func_array($columnDef['method'], [$whereQuery, $globalKeyword]);
+                            call_user_func_array($columnDef['method'], [$whereQuery, $keyword]);
                             $queryBuilder->addNestedWhereQuery($whereQuery, 'or');
                         } else {
                             $this->compileColumnQuery(
@@ -156,7 +188,7 @@ class QueryBuilderEngine extends BaseEngine
                                 Helper::getOrMethod($columnDef['method']),
                                 $columnDef['parameters'],
                                 $columnName,
-                                $globalKeyword
+                                $keyword
                             );
                         }
                     } else {
@@ -170,13 +202,13 @@ class QueryBuilderEngine extends BaseEngine
                                     $queryBuilder,
                                     $relation,
                                     $relationColumn,
-                                    $globalKeyword
+                                    $keyword
                                 );
                             } else {
-                                $this->compileQuerySearch($queryBuilder, $columnName, $globalKeyword);
+                                $this->compileQuerySearch($queryBuilder, $columnName, $keyword);
                             }
                         } else {
-                            $this->compileQuerySearch($queryBuilder, $columnName, $globalKeyword);
+                            $this->compileQuerySearch($queryBuilder, $columnName, $keyword);
                         }
                     }
 
@@ -184,6 +216,17 @@ class QueryBuilderEngine extends BaseEngine
                 }
             }
         );
+    }
+
+    /**
+     * Check if column has custom filter handler.
+     *
+     * @param  string $columnName
+     * @return bool
+     */
+    public function hasCustomFilter($columnName)
+    {
+        return isset($this->columnDef['filter'][$columnName]);
     }
 
     /**
@@ -326,10 +369,10 @@ class QueryBuilderEngine extends BaseEngine
          */
         foreach ($relationChunk as $relation => $chunk) {
             // Prepare variables
-            $builder      = $chunk['builder'];
-            $query        = $chunk['query'];
-            $bindings     = $builder->getBindings();
-            $builder      = "({$builder->toSql()}) >= 1";
+            $builder  = $chunk['builder'];
+            $query    = $chunk['query'];
+            $bindings = $builder->getBindings();
+            $builder  = "({$builder->toSql()}) >= 1";
 
             // Check if it last relation we will use orWhereRaw
             if ($lastRelation == $relation) {
@@ -513,42 +556,52 @@ class QueryBuilderEngine extends BaseEngine
      */
     protected function joinEagerLoadedColumn($relation, $relationColumn)
     {
-        $model = $this->query->getRelation($relation);
-        switch (true) {
-            case $model instanceof BelongsToMany:
-                $pivot   = $model->getTable();
-                $pivotPK = $model->getExistenceCompareKey();
-                $pivotFK = $model->getQualifiedParentKeyName();
-                $this->performJoin($pivot, $pivotPK, $pivotFK);
+        $lastQuery = $this->query;
+        foreach (explode('.', $relation) as $eachRelation) {
+            $model = $lastQuery->getRelation($eachRelation);
+            switch (true) {
+                case $model instanceof BelongsToMany:
+                    $pivot   = $model->getTable();
+                    $pivotPK = $model->getExistenceCompareKey();
+                    $pivotFK = $model->getQualifiedParentKeyName();
+                    $this->performJoin($pivot, $pivotPK, $pivotFK);
 
-                $related = $model->getRelated();
-                $table   = $related->getTable();
-                $tablePK = $related->getForeignKey();
-                $foreign = $pivot . '.' . $tablePK;
-                $other   = $related->getQualifiedKeyName();
+                    $related = $model->getRelated();
+                    $table   = $related->getTable();
+                    $tablePK = $related->getForeignKey();
+                    $foreign = $pivot . '.' . $tablePK;
+                    $other   = $related->getQualifiedKeyName();
 
-                $this->query->addSelect($table . '.' . $relationColumn);
-                break;
+                    $lastQuery->addSelect($table . '.' . $relationColumn);
+                    $this->performJoin($table, $foreign, $other);
 
-            case $model instanceof HasOneOrMany:
-                $table   = $model->getRelated()->getTable();
-                $foreign = $model->getQualifiedForeignKeyName();
-                $other   = $model->getQualifiedParentKeyName();
-                break;
+                    break;
 
-            case $model instanceof BelongsTo:
-                $table   = $model->getRelated()->getTable();
-                $foreign = $model->getQualifiedForeignKey();
-                $other   = $model->getQualifiedOwnerKeyName();
-                break;
+                case $model instanceof HasOneOrMany:
+                    $table   = $model->getRelated()->getTable();
+                    $foreign = $model->getQualifiedForeignKeyName();
+                    $other   = $model->getQualifiedParentKeyName();
+                    break;
 
-            default:
-                $table   = $model->getRelated()->getTable();
-                $foreign = $model->getQualifiedForeignKey();
-                $other   = $model->getQualifiedOtherKeyName();
+                case $model instanceof BelongsTo:
+                    $table   = $model->getRelated()->getTable();
+                    $foreign = $model->getQualifiedForeignKey();
+                    $other   = $model->getQualifiedOwnerKeyName();
+                    break;
+
+                default:
+                    $table = $model->getRelated()->getTable();
+                    if ($model instanceof HasOneOrMany) {
+                        $foreign = $model->getForeignKey();
+                        $other   = $model->getQualifiedParentKeyName();
+                    } else {
+                        $foreign = $model->getQualifiedForeignKey();
+                        $other   = $model->getQualifiedOtherKeyName();
+                    }
+            }
+            $this->performJoin($table, $foreign, $other);
+            $lastQuery = $model->getQuery();
         }
-
-        $this->performJoin($table, $foreign, $other);
 
         return $table . '.' . $relationColumn;
     }
@@ -625,11 +678,11 @@ class QueryBuilderEngine extends BaseEngine
         foreach ($this->request->orderableColumns() as $orderable) {
             $column = $this->getColumnName($orderable['column'], true);
 
-            if ($this->isBlacklisted($column)) {
+            if ($this->isBlacklisted($column) && ! $this->hasCustomOrder($column)) {
                 continue;
             }
 
-            if (isset($this->columnDef['order'][$column])) {
+            if ($this->hasCustomOrder($column)) {
                 $method     = $this->columnDef['order'][$column]['method'];
                 $parameters = $this->columnDef['order'][$column]['parameters'];
                 $this->compileColumnQuery(
@@ -648,8 +701,20 @@ class QueryBuilderEngine extends BaseEngine
                     $relation       = implode('.', $parts);
 
                     if (in_array($relation, $eagerLoads)) {
-                        $relationship = $this->query->getRelation($relation);
-                        if (! ($relationship instanceof MorphToMany)) {
+                        // Loop for nested relations
+                        // This code is check morph many or not.
+                        // If one of nested relation is MorphToMany
+                        // we will call joinEagerLoadedColumn.
+                        $lastQuery     = $this->query;
+                        $isMorphToMany = false;
+                        foreach (explode('.', $relation) as $eachRelation) {
+                            $relationship = $lastQuery->getRelation($eachRelation);
+                            if (! ($relationship instanceof MorphToMany)) {
+                                $isMorphToMany = true;
+                            }
+                            $lastQuery = $relationship;
+                        }
+                        if ($isMorphToMany) {
                             $column = $this->joinEagerLoadedColumn($relation, $relationColumn);
                         } else {
                             $valid = 0;
@@ -666,6 +731,17 @@ class QueryBuilderEngine extends BaseEngine
                 }
             }
         }
+    }
+
+    /**
+     * Check if column has custom sort handler.
+     *
+     * @param string $column
+     * @return bool
+     */
+    protected function hasCustomOrder($column)
+    {
+        return isset($this->columnDef['order'][$column]);
     }
 
     /**
@@ -701,5 +777,20 @@ class QueryBuilderEngine extends BaseEngine
     public function results()
     {
         return $this->query->get();
+    }
+
+    /**
+     * Add column in collection.
+     *
+     * @param string $name
+     * @param string|callable $content
+     * @param bool|int $order
+     * @return $this
+     */
+    public function addColumn($name, $content, $order = false)
+    {
+        $this->pushToBlacklist($name);
+
+        return parent::addColumn($name, $content, $order);
     }
 }