+++ tests/Integration/QueryBuilderEngineTest.php
@@ -3,6 +3,8 @@
 namespace Yajra\Datatables\Tests\Integration;
 
 use DB;
+use Illuminate\Database\Query\Builder;
+use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Http\JsonResponse;
 use Yajra\Datatables\Datatables;
 use Yajra\Datatables\Engines\QueryBuilderEngine;
@@ -11,6 +13,8 @@ use Yajra\Datatables\Tests\TestCase;
 
 class QueryBuilderEngineTest extends TestCase
 {
+    use DatabaseTransactions;
+
     /** @test */
     public function it_returns_all_records_when_no_parameters_is_passed()
     {
@@ -30,7 +34,25 @@ class QueryBuilderEngineTest extends TestCase
                 ['data' => 'name', 'name' => 'name', 'searchable' => "true", 'orderable' => "true"],
                 ['data' => 'email', 'name' => 'email', 'searchable' => "true", 'orderable' => "true"],
             ],
-            'search'  => ['value' => 'Record 19'],
+            'search'  => ['value' => 'Record-19'],
+        ]);
+
+        $crawler->assertJson([
+            'draw'            => 0,
+            'recordsTotal'    => 20,
+            'recordsFiltered' => 1,
+        ]);
+    }
+
+    /** @test */
+    public function it_can_perform_multiple_term_global_search()
+    {
+        $crawler = $this->call('GET', '/queryBuilder/users', [
+            'columns' => [
+                ['data' => 'name', 'name' => 'name', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'email', 'name' => 'email', 'searchable' => "true", 'orderable' => "true"],
+            ],
+            'search'  => ['value' => 'Record-19 Email-19'],
         ]);
 
         $crawler->assertJson([
@@ -43,8 +65,7 @@ class QueryBuilderEngineTest extends TestCase
     /** @test */
     public function it_accepts_a_query_builder_using_of_factory()
     {
-        $dataTable = Datatables::of(DB::table('users'
-        ));
+        $dataTable = Datatables::of(DB::table('users'));
         $response  = $dataTable->make(true);
         $this->assertInstanceOf(QueryBuilderEngine::class, $dataTable);
         $this->assertInstanceOf(JsonResponse::class, $response);
@@ -53,8 +74,7 @@ class QueryBuilderEngineTest extends TestCase
     /** @test */
     public function it_accepts_a_query_builder_using_facade()
     {
-        $dataTable = DatatablesFacade::of(DB::table('users'
-        ));
+        $dataTable = DatatablesFacade::of(DB::table('users'));
         $response  = $dataTable->make(true);
         $this->assertInstanceOf(QueryBuilderEngine::class, $dataTable);
         $this->assertInstanceOf(JsonResponse::class, $response);
@@ -63,8 +83,7 @@ class QueryBuilderEngineTest extends TestCase
     /** @test */
     public function it_accepts_a_query_builder_using_facade_queryBuilder_method()
     {
-        $dataTable = DatatablesFacade::queryBuilder(DB::table('users'
-        ));
+        $dataTable = DatatablesFacade::queryBuilder(DB::table('users'));
         $response  = $dataTable->make(true);
         $this->assertInstanceOf(QueryBuilderEngine::class, $dataTable);
         $this->assertInstanceOf(JsonResponse::class, $response);
@@ -82,19 +101,101 @@ class QueryBuilderEngineTest extends TestCase
     /** @test */
     public function it_accepts_a_query_builder_using_ioc_container_factory()
     {
-        $dataTable = app('datatables')->of(DB::table('users'
-        ));
+        $dataTable = app('datatables')->of(DB::table('users'));
         $response  = $dataTable->make(true);
         $this->assertInstanceOf(QueryBuilderEngine::class, $dataTable);
         $this->assertInstanceOf(JsonResponse::class, $response);
+    }
+
+    /** @test */
+    public function it_does_not_allow_search_on_added_columns()
+    {
+        $crawler = $this->call('GET', '/queryBuilder/addColumn', [
+            'columns' => [
+                ['data' => 'foo', 'name' => 'foo', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'name', 'name' => 'name', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'email', 'name' => 'email', 'searchable' => "true", 'orderable' => "true"],
+            ],
+            'search'  => ['value' => 'Record-19'],
+        ]);
+
+        $crawler->assertJson([
+            'draw'            => 0,
+            'recordsTotal'    => 20,
+            'recordsFiltered' => 1,
+        ]);
+    }
+
+    /** @test */
+    public function it_can_return_auto_index_column()
+    {
+        $crawler = $this->call('GET', '/queryBuilder/indexColumn', [
+            'columns' => [
+                ['data' => 'DT_Row_index', 'name' => 'index', 'searchable' => "false", 'orderable' => "false"],
+                ['data' => 'name', 'name' => 'name', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'email', 'name' => 'email', 'searchable' => "true", 'orderable' => "true"],
+            ],
+            'search'  => ['value' => 'Record-19'],
+        ]);
+
+        $crawler->assertJson([
+            'draw'            => 0,
+            'recordsTotal'    => 20,
+            'recordsFiltered' => 1,
+        ]);
+
+        $this->assertArrayHasKey('DT_Row_Index', $crawler->json()['data'][0]);
+    }
+
+    /** @test */
+    public function it_allows_search_on_added_column_with_custom_filter_handler()
+    {
+        $crawler = $this->call('GET', '/queryBuilder/filterColumn', [
+            'columns' => [
+                ['data' => 'foo', 'name' => 'foo', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'name', 'name' => 'name', 'searchable' => "true", 'orderable' => "true"],
+                ['data' => 'email', 'name' => 'email', 'searchable' => "true", 'orderable' => "true"],
+            ],
+            'search'  => ['value' => 'Record-19'],
+        ]);
+
+        $crawler->assertJson([
+            'draw'            => 0,
+            'recordsTotal'    => 20,
+            'recordsFiltered' => 1,
+        ]);
+
+        $queries = $crawler->json()['queries'];
+        $this->assertContains('"1" = ?', $queries[1]['query']);
     }
 
     protected function setUp()
     {
         parent::setUp();
 
-        $this->app['router']->get('/queryBuilder/users', function (Datatables $datatables) {
-            return $datatables->queryBuilder(DB::table('users'))->make('true');
+        $this->app['router']->get('/queryBuilder/users', function (Datatables $dataTable) {
+            return $dataTable->queryBuilder(DB::table('users'))->make('true');
+        });
+
+        $this->app['router']->get('/queryBuilder/addColumn', function (Datatables $dataTable) {
+            return $dataTable->queryBuilder(DB::table('users'))
+                             ->addColumn('foo', 'bar')
+                             ->make('true');
+        });
+
+        $this->app['router']->get('/queryBuilder/indexColumn', function (Datatables $dataTable) {
+            return $dataTable->queryBuilder(DB::table('users'))
+                             ->addIndexColumn()
+                             ->make('true');
+        });
+
+        $this->app['router']->get('/queryBuilder/filterColumn', function (Datatables $dataTable) {
+            return $dataTable->queryBuilder(DB::table('users'))
+                             ->addColumn('foo', 'bar')
+                             ->filterColumn('foo', function (Builder $builder, $keyword) {
+                                 $builder->where('1', $keyword);
+                             })
+                             ->make('true');
         });
     }
 }